Implement ElasticSearch Synchronization Handlers

+++ MicroStore.BuildingBlocks.Paging/PagedResult.cs
@@ -4,14 +4,14 @@
     {
         public int Lenght { get; protected set; }
         public int Skip { get; set; }
-        public int TotalCount { get; protected set; }
+        public long TotalCount { get; protected set; }
     }
 
     public class PagedResult<T> : PagedResult 
     { 
         public IEnumerable<T> Items { get; set; }
 
-        public PagedResult(IEnumerable<T> items, int count, int skip, int lenght)
+        public PagedResult(IEnumerable<T> items, long count, int skip, int lenght)
         {
             Items = items;
             TotalCount = count;
